Refactor code and created methods for DB
